src/requirements.txt
async-timeout==4.0.3
azure-core==1.30.2
azure-keyvault==4.2.0
azure-identity==1.17.1
certifi==2024.7.4
python-dotenv==1.0.1
regex==2024.5.15
streamlit==1.37.0
asyncio==3.4.3
azure-storage-blob==12.19.0
aiohttp==3.10.2
streamlit_authenticator==0.3.3